app/backend/zamow_jedzenie/requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
gunicorn==20.1.0
django-cors-headers==3.11.0
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.1.0
django-filter==21.1
